git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@14638 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/network/protocols/connect_to_peer.cpp
@@ -46,13 +46,14 @@ ConnectToPeer::~ConnectToPeer()
 
 // ----------------------------------------------------------------------------
 
-void ConnectToPeer::notifyEvent(Event* event)
+bool ConnectToPeer::notifyEventAsynchronous(Event* event)
 {
     if (event->type == EVENT_TYPE_CONNECTED)
     {
         Log::debug("ConnectToPeer", "Received event notifying peer connection.");
         m_state = CONNECTED; // we received a message, we are connected
     }
+    return true;
 }
 
 // ----------------------------------------------------------------------------
@@ -85,8 +86,28 @@ void ConnectToPeer::asynchronousUpdate()
             {
                 if (m_peer_address.ip != 0 && m_peer_address.port != 0)
                 {
+                    // we're in the same lan (same public ip address) !!
+                    if (m_peer_address.ip == NetworkManager::getInstance()->getPublicAddress().ip)
+                    {
+                        // just send a broadcast packet with the string aloha_stk inside, the client will know our ip address and will connect
+                        STKHost* host = NetworkManager::getInstance()->getHost();
+                        TransportAddress broadcast_address;
+                        broadcast_address.ip = -1; // 255.255.255.255
+                        broadcast_address.port = m_peer_address.port; // 0b10101100000101101101111111111111; // for test
+                        char data[] = "aloha_stk\0";
+                        host->sendRawPacket((uint8_t*)(data), 10, broadcast_address);
+                        Log::info("ConnectToPeer", "Broadcast aloha sent.");
+                        StkTime::sleep(1);
+                        broadcast_address.ip = 0x7f000001; // 127.0.0.1 (localhost)
+                        broadcast_address.port = m_peer_address.port;
+                        host->sendRawPacket((uint8_t*)(data), 10, broadcast_address);
+                        Log::info("ConnectToPeer", "Broadcast aloha to self.");
+                    }
+                    else
+                    {
+                        m_current_protocol_id = m_listener->requestStart(new PingProtocol(m_peer_address, 2.0));
+                    }
                     m_state = CONNECTING;
-                    m_current_protocol_id = m_listener->requestStart(new PingProtocol(m_peer_address, 2.0));
                 }
                 else
                 {
@@ -96,9 +117,12 @@ void ConnectToPeer::asynchronousUpdate()
             }
             break;
         case CONNECTING: // waiting the peer to connect
+            break;
         case CONNECTED:
         {
-            m_listener->requestTerminate( m_listener->getProtocol(m_current_protocol_id)); // kill the ping protocol because we're connected
+            // the ping protocol is there for NAT traversal (disabled when connecting to LAN peer)
+            if (m_peer_address != NetworkManager::getInstance()->getPublicAddress())
+                m_listener->requestTerminate( m_listener->getProtocol(m_current_protocol_id)); // kill the ping protocol because we're connected
             m_state = DONE;
             break;
         }
